requirements.txt
opencv-python-headless
streamlit
pytube
ultralytics-yolov8
